+++ host/lib/usrp/multi_crimson.cpp
@@ -461,7 +461,7 @@ void multi_crimson::set_rx_rate(double rate, size_t chan){
     double actual_rate = _tree->access<double>(rx_dsp_root(chan) / "rate" / "value").get();
 
     // re-tune the frequency
-    int cur_dsp_nco = _tree->access<int>(rx_dsp_root(chan) / "nco").get();
+    double cur_dsp_nco = _tree->access<double>(rx_dsp_root(chan) / "nco").get();
     double cur_lo_freq = 0;
     if (_tree->access<int>(rx_rf_fe_root(chan) / "freq" / "band").get() == 1) {
         cur_lo_freq = _tree->access<double>(rx_rf_fe_root(chan) / "freq" / "value").get();
@@ -511,9 +511,9 @@ tune_result_t multi_crimson::set_rx_freq(const tune_request_t &tune_request, siz
     if (*freq > 15000000.0 && !(cur_rx_rate > (CRIMSON_MASTER_CLOCK_RATE / 9)) ) {
        *freq -= 15000000.0;
        offset = true;
-       _tree->access<int>(rx_dsp_root(chan) / "nco").set(-15000000);
+       _tree->access<double>(rx_dsp_root(chan) / "nco").set(-15000000);
     } else {
-       _tree->access<int>(rx_dsp_root(chan) / "nco").set(0);
+       _tree->access<double>(rx_dsp_root(chan) / "nco").set(0);
     }
 
     // check the tuning ranges first, and clip if necessary
@@ -528,9 +528,9 @@ tune_result_t multi_crimson::set_rx_freq(const tune_request_t &tune_request, siz
 
     // use the DSP NCO with base band
     if (_tree->access<int>(rx_rf_fe_root(chan) / "freq" / "band").get() == 0) {
-        int cur_dsp_nco = _tree->access<int>(rx_dsp_root(chan) / "nco").get();
-	int set_dsp_nco = cur_dsp_nco - *freq;
-        _tree->access<int>(rx_dsp_root(chan) / "nco").set(set_dsp_nco);
+        double cur_dsp_nco = _tree->access<double>(rx_dsp_root(chan) / "nco").get();
+        double set_dsp_nco = cur_dsp_nco - *freq;
+        _tree->access<double>(rx_dsp_root(chan) / "nco").set(set_dsp_nco);
 
 	result.actual_rf_freq = -set_dsp_nco;
 
@@ -540,14 +540,14 @@ tune_result_t multi_crimson::set_rx_freq(const tune_request_t &tune_request, siz
 
 	// read back the frequency and adjust for the errors with DSP NCO if possible
         double cur_lo_freq = _tree->access<double>(rx_rf_fe_root(chan) / "freq" / "value").get();
-        int cur_dsp_nco = _tree->access<int>(rx_dsp_root(chan) / "nco").get();
-	int set_dsp_nco = cur_dsp_nco - (*freq - cur_lo_freq);
+        double cur_dsp_nco = _tree->access<double>(rx_dsp_root(chan) / "nco").get();
+        double set_dsp_nco = cur_dsp_nco - (*freq - cur_lo_freq);
 
-	if (set_dsp_nco >  161000000) set_dsp_nco = 161000000;
-	if (set_dsp_nco < -161000000) set_dsp_nco = -161000000;
-        _tree->access<int>(rx_dsp_root(chan) / "nco").set(set_dsp_nco);
+        if (set_dsp_nco >  161000000) set_dsp_nco = 161000000;
+        if (set_dsp_nco < -161000000) set_dsp_nco = -161000000;
+        _tree->access<double>(rx_dsp_root(chan) / "nco").set(set_dsp_nco);
 
-	result.actual_rf_freq = cur_lo_freq - set_dsp_nco;
+        result.actual_rf_freq = cur_lo_freq - set_dsp_nco;
     }
 
     // account back for the offset
@@ -608,7 +608,7 @@ tune_result_t multi_crimson::set_rx_freq(const tune_request_t &tune_request, siz
 
 // get the RX frequency on specified channel
 double multi_crimson::get_rx_freq(size_t chan){
-    int cur_dsp_nco = _tree->access<int>(rx_dsp_root(chan) / "nco").get();
+    double cur_dsp_nco = _tree->access<double>(rx_dsp_root(chan) / "nco").get();
     double cur_lo_freq = 0;
     if (_tree->access<int>(rx_rf_fe_root(chan) / "freq" / "band").get() == 1) {
         cur_lo_freq = _tree->access<double>(rx_rf_fe_root(chan) / "freq" / "value").get();
@@ -735,8 +735,8 @@ void multi_crimson::set_tx_rate(double rate, size_t chan){
     double actual_rate = _tree->access<double>(tx_dsp_root(chan) / "rate" / "value").get();
 
     // re-tune the frequency
-    int cur_dac_nco = _tree->access<int>(tx_rf_fe_root(chan) / "nco").get();
-    int cur_dsp_nco = _tree->access<int>(tx_dsp_root(chan) / "nco").get();
+    double cur_dac_nco = _tree->access<double>(tx_rf_fe_root(chan) / "nco").get();
+    double cur_dsp_nco = _tree->access<double>(tx_dsp_root(chan) / "nco").get();
     double cur_lo_freq = 0;
     if (_tree->access<int>(tx_rf_fe_root(chan) / "freq" / "band").get() == 1) {
     	cur_lo_freq = _tree->access<double>(tx_rf_fe_root(chan) / "freq" / "value").get();
@@ -786,9 +786,9 @@ tune_result_t multi_crimson::set_tx_freq(const tune_request_t &tune_request, siz
     if (*freq > 85000000.0 && !(cur_tx_rate > (CRIMSON_MASTER_CLOCK_RATE / 9))) {
        *freq -= 85000000.0;
        offset = true;
-       _tree->access<int>(tx_rf_fe_root(chan) / "nco").set(85000000);
+       _tree->access<double>(tx_rf_fe_root(chan) / "nco").set(85000000);
     } else {
-       _tree->access<int>(tx_rf_fe_root(chan) / "nco").set(0);
+       _tree->access<double>(tx_rf_fe_root(chan) / "nco").set(0);
     }
 
     // check the tuning ranges first, and clip if necessary
@@ -803,11 +803,11 @@ tune_result_t multi_crimson::set_tx_freq(const tune_request_t &tune_request, siz
 
     // use the DSP NCO with base band
     if (_tree->access<int>(tx_rf_fe_root(chan) / "freq" / "band").get() == 0) {
-        int cur_dac_nco = _tree->access<int>(tx_rf_fe_root(chan) / "nco").get();
-	int set_dsp_nco = *freq;
-        _tree->access<int>(tx_dsp_root(chan) / "nco").set(set_dsp_nco);
+        double cur_dac_nco = _tree->access<double>(tx_rf_fe_root(chan) / "nco").get();
+        double set_dsp_nco = *freq;
+        _tree->access<double>(tx_dsp_root(chan) / "nco").set(set_dsp_nco);
 
-	result.actual_rf_freq = set_dsp_nco + cur_dac_nco;
+        result.actual_rf_freq = set_dsp_nco + cur_dac_nco;
 
     // use the LO with high band
     } else {
@@ -815,14 +815,14 @@ tune_result_t multi_crimson::set_tx_freq(const tune_request_t &tune_request, siz
 
 	// read back the frequency and adjust for the errors with DSP NCO if possible
         double cur_lo_freq = _tree->access<double>(tx_rf_fe_root(chan) / "freq" / "value").get();
-        int cur_dac_nco = _tree->access<int>(tx_rf_fe_root(chan) / "nco").get();
-	int set_dsp_nco = *freq - cur_lo_freq;
+        double cur_dac_nco = _tree->access<double>(tx_rf_fe_root(chan) / "nco").get();
+        int set_dsp_nco = *freq - cur_lo_freq;
 
-	if (set_dsp_nco >  161000000) set_dsp_nco = 161000000;
-	if (set_dsp_nco < -161000000) set_dsp_nco = -161000000;
-        _tree->access<int>(tx_dsp_root(chan) / "nco").set(set_dsp_nco);
+        if (set_dsp_nco >  161000000) set_dsp_nco = 161000000;
+        if (set_dsp_nco < -161000000) set_dsp_nco = -161000000;
+        _tree->access<double>(tx_dsp_root(chan) / "nco").set(set_dsp_nco);
 
-	result.actual_rf_freq = cur_lo_freq + set_dsp_nco + cur_dac_nco;
+        result.actual_rf_freq = cur_lo_freq + set_dsp_nco + cur_dac_nco;
     }
 
     // account back for the offset
@@ -883,8 +883,8 @@ tune_result_t multi_crimson::set_tx_freq(const tune_request_t &tune_request, siz
 
 // get the TX frequency on specified channel
 double multi_crimson::get_tx_freq(size_t chan){
-    int cur_dac_nco = _tree->access<int>(tx_rf_fe_root(chan) / "nco").get();
-    int cur_dsp_nco = _tree->access<int>(tx_dsp_root(chan) / "nco").get();
+    double cur_dac_nco = _tree->access<double>(tx_rf_fe_root(chan) / "nco").get();
+    double cur_dsp_nco = _tree->access<double>(tx_dsp_root(chan) / "nco").get();
     double cur_lo_freq = 0;
     if (_tree->access<int>(tx_rf_fe_root(chan) / "freq" / "band").get() == 1) {
     	cur_lo_freq = _tree->access<double>(tx_rf_fe_root(chan) / "freq" / "value").get();
